Created new UI definition for time selector popup and new empty python class for this popup
Buttons in this class are connected to empty signals which can become connected to some other function.
The class takes a parent widget as attribute; which could be updated …
The class also takes the start and end times as hours, minutes and seconds; the spin buttons are initialized with these values.